Function openrpc_types::resolve_within

pub fn resolve_within(openrpc: OpenRPC) -> Result<OpenRPC, BrokenReference>

Try and resolve all ReferenceOrs in the document by looking up references in OpenRPC::components.

Returns a BrokenReference on the first ReferenceOr::Reference with no corresponding entry in OpenRPC::components.

Note that unlike schemars::schema::SchemaObject::references, OpenRPC’s references are acyclic, so this is guaranteed to terminate.